+++ checkpoint/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow ArrayMetadataStore to update existing metadatas.
 
+### Added
+
+- Orbax Emergency Checkpointer will now do single slice restore from persistent storage.
+
 ## [0.11.21] - 2025-08-07
 
 ### Fixed